requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==4.1.13
django-bootstrap-v5==1.0.11
idna==3.10
oauthlib==3.3.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.0.0
social-auth-core==4.6.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
djangorestframework==3.14.0